+++ src/main/java/com/react/pnld/dto/FileResumeDTO.java
@@ -1,4 +1,4 @@
-package com.react.pnld.model.dto;
+package com.react.pnld.dto;
 
 public class FileResumeDTO {
     private int totalRecords;
